+++ composeApp/src/commonMain/kotlin/org/hank/botm/data/database/dao/ResultDao.kt
@@ -14,7 +14,7 @@ interface ResultDao {
     fun getAllResults(): Flow<List<ResultEntity>>
 
     @Query("SELECT * FROM result WHERE roundId = :roundId")
-    fun getRoundResults(roundId: Long): Flow<List<ResultEntity>>
+    fun getRoundResults(roundId: Int): Flow<List<ResultEntity>>
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insertResults(result: List<ResultEntity>): List<Long>
