+++ composeApp/src/commonMain/kotlin/org/hank/botm/data/repository/PlayerRepository.kt
@@ -10,19 +10,19 @@ import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.withContext
 
 interface PlayerRepository {
-    fun getPlayers(gameId: Long): Flow<List<Player>>
+    fun getPlayers(gameId: Int): Flow<List<Player>>
     fun getCurrentGamePlayers(): Flow<List<Player>>
 
     suspend fun insertPlayer(player: Player)
-    suspend fun getPlayer(id: Long): Player
-    suspend fun updatePlayerBalance(id: Long, balance: Int)
+    suspend fun getPlayer(id: Int): Player
+    suspend fun updatePlayerBalance(id: Int, balance: Int)
 }
 
 class PlayerRepositoryImpl constructor(
     private val playerDao: PlayerDao,
     private val ioDispatcher: CoroutineDispatcher
 ) : PlayerRepository {
-    override fun getPlayers(gameId: Long): Flow<List<Player>> {
+    override fun getPlayers(gameId: Int): Flow<List<Player>> {
         return playerDao.getPlayersByGame(gameId).map {
             it.map { it.asDomain() }
         }
@@ -38,11 +38,11 @@ class PlayerRepositoryImpl constructor(
         playerDao.insertPlayer(player.asEntity())
     }
 
-    override suspend fun getPlayer(id: Long): Player = withContext(ioDispatcher) {
+    override suspend fun getPlayer(id: Int): Player = withContext(ioDispatcher) {
         playerDao.getPlayer(id).asDomain()
     }
 
-    override suspend fun updatePlayerBalance(id: Long, profit: Int) {
+    override suspend fun updatePlayerBalance(id: Int, profit: Int) {
         playerDao.updatePlayerBalance(id, profit)
     }
 }
