+++ composeApp/src/commonMain/kotlin/org/hank/botm/data/repository/ResultRepository.kt
@@ -11,8 +11,8 @@ import kotlinx.coroutines.withContext
 
 interface ResultRepository {
     fun getAllResults(): Flow<List<Result>>
-    fun getRoundResults(roundId: Long): Flow<List<Result>>
-    suspend fun insertResults(results: List<Result>, roundId: Long)
+    fun getRoundResults(roundId: Int): Flow<List<Result>>
+    suspend fun insertResults(results: List<Result>, roundId: Int)
 }
 
 class ResultRepositoryImpl (
@@ -27,13 +27,13 @@ class ResultRepositoryImpl (
         }
     }
 
-    override fun getRoundResults(roundId: Long): Flow<List<Result>> =
+    override fun getRoundResults(roundId: Int): Flow<List<Result>> =
         resultDao.getRoundResults(roundId).map {
             it.map { it.asDomain() }
         }
 
 
-    override suspend fun insertResults(results: List<Result>, roundId: Long) = withContext(ioDispatcher) {
+    override suspend fun insertResults(results: List<Result>, roundId: Int) = withContext(ioDispatcher) {
         resultDao.insertResults(results.map { it.asEntity(roundId) })
         Unit
     }
