+++ composeApp/src/commonMain/kotlin/org/hank/botm/domain/model/Result.kt
@@ -4,20 +4,19 @@ import org.hank.botm.data.database.model.ResultEntity
 import org.hank.botm.data.network.model.PlayerResultDto
 
 data class Result(
-    val id: Long = 0,
-    val roundId: Long = 0,
-    val playerId: Long,
+    val id: Int = 0,
+    val roundId: Int = 0,
+    val playerId: Int,
     val profit: Int = 0,
 )
 
-fun Result.asEntity(roundId: Long) = ResultEntity(
+fun Result.asEntity(roundId: Int) = ResultEntity(
     roundId = roundId,
     playerId = playerId,
     profit = profit
 )
 
 fun Result.asPlayerResultDto() = PlayerResultDto(
-    // FIXME("Change app's id to Int")
-    playerId = playerId.toInt(),
+    playerId = playerId,
     profit = profit
 )
