+++ composeApp/src/commonMain/kotlin/org/hank/botm/domain/model/Round.kt
@@ -3,9 +3,9 @@ package org.hank.botm.domain.model
 import org.hank.botm.data.database.model.RoundEntity
 
 data class Round(
-    val id: Long = 0,
+    val id: Int = 0,
     val bet: Int,
-    val gameId: Long,
+    val gameId: Int,
 )
 
 // We set the RoundEntity id to 0 for autoGenerate the primary key.
