+++ composeApp/src/commonMain/kotlin/org/hank/botm/ui/model/PlayerResult.kt
@@ -2,7 +2,7 @@ package org.hank.botm.ui.model
 
 data class PlayerResult(
     val playerName: String,
-    val playerId: Long,
+    val playerId: Int,
     val cardsInHand: String = "",
     val isWinner: Boolean = false,
 )
