+++ composeApp/src/commonMain/kotlin/org/hank/botm/ui/model/RoundResult.kt
@@ -3,6 +3,6 @@ package org.hank.botm.ui.model
 import org.hank.botm.domain.model.Result
 
 data class RoundResult(
-    val roundId: Long,
+    val roundId: Int,
     val results: List<Result>,
 )
